Update README.md
In this Machine Learning (ML) End-to-End Project, a complete ML pipeline was developed. "Heart_Disease_Dataset" was utilized, which is available on the Kaggle. The project was developed in five main parts:
1.	Data cleaning
2.	Model building
3.	Model saving and loading
4.	FastAPI endpoint for the model
5.	Deployment of the FastAPI on Hugging Face Space, by following steps:-
(5.1) Deployment of the FastAPI application on Hugging Face and creation of a simple user interface (UI) for the FastAPI using Gradio or Streamlit application that interacts with the FastAPI endpoint.
(5.2)  Deployment of the Gradioor or Streamlit script (app.py) on Hugging Face Spaces.